Use double for storing the average timestamps per frame.
This fixes a number of imprecisions around timing.
In several places we were calculating a target timestamp or duration in timestamps based on a number of frames multiplied by this value.
In many videos the combination of timebase and framerate makes this number non-integer which was causing large and cumulative error.

Other changes as part of this refactoring:
- When converting to text-based timecode the result is now rounded to the nearest centisecond (before: truncated).
- When converting to numerical times like seconds, centi-minutes or milli-hours, the time is no longer rounded to the nearest millisecond.

+++ Kinovea.ScreenManager/Metadata/DrawingToolbar/DrawingTool.cs
@@ -103,11 +103,11 @@ namespace Kinovea.ScreenManager
                 this.currentStyleElements = defaultStyle.Clone();
         }
 
-        public override AbstractDrawing GetNewDrawing(PointF origin, long timestamp, long averageTimeStampsPerFrame, IImageToViewportTransformer transformer)
+        public override AbstractDrawing GetNewDrawing(PointF origin, long timestamp, double averageTimeStampsPerFrame, IImageToViewportTransformer transformer)
         {
             // Drawings constructors must conforms to one of two predefined prototypes:
-            // (PointF origin, long timestamp, long averageTimeStampsPerFrame, DrawingStyle preset = null, IImageToViewportTransformer transformer = null).
-            // (PointF origin, long timestamp, long averageTimeStampsPerFrame, DrawingStyle stylePreset).
+            // (PointF origin, long timestamp, double averageTimeStampsPerFrame, DrawingStyle preset = null, IImageToViewportTransformer transformer = null).
+            // (PointF origin, long timestamp, double averageTimeStampsPerFrame, DrawingStyle stylePreset).
             // Exemple for 1: DrawingAngle.
             // Exemple for 2: DrawingText.
