+++ Kinovea.ScreenManager/Metadata/Drawings/DrawingLine.cs
@@ -131,7 +131,7 @@ namespace Kinovea.ScreenManager
         #endregion
 
         #region Constructors
-        public DrawingLine(PointF origin, long timestamp, long averageTimeStampsPerFrame, StyleElements preset = null, IImageToViewportTransformer transformer = null)
+        public DrawingLine(PointF origin, long timestamp, double averageTimeStampsPerFrame, StyleElements preset = null, IImageToViewportTransformer transformer = null)
         {
             points["a"] = origin;
             points["b"] = origin.Translate(10, 0);
